+++ lib/blackmagicdesign.dart
@@ -3,3 +3,4 @@ library blackmagicdesign;
 
 export 'atem/atem.dart';
 export 'hyperdeck/hyperdeck.dart';
+export 'web_presenter/web_presenter.dart';